requirements.txt
Flask>=2.1
openai>=0.27.0
reportlab>=4.0
python-docx>=0.8.11
gunicorn>=20.1.0
MarkupSafe>=2.1.1
